+++ CHANGELOG.md
@@ -22,3 +22,12 @@
 # 1.1.3
 
 - Added configurable to set night vision on or off when game loads.
+
+# 2.0.0
+
+- Fixed bug where night vision would not be enabled by default when configured to do so
+- Reordered config file, you may need to delete the old one and generate a new one
+- Night vision now works outside!
+- Added option to toggle on and off fog
+- The fog toggle may also be affect steam valve indoors, might add an option to disable that feature
+- May possibly conflict with HDLethalCompany fog settings